+++ config.js
@@ -1,0 +1,4 @@
+module.exports = {
+    PORT: process.env.PORT || 3000,
+    MONGO_URL: 'mongodb://localhost:27017/blog'
+};
