+++ npm-debug.log
@@ -1,0 +1,45 @@
+0 info it worked if it ends with ok
+1 verbose cli [ '/usr/bin/node', '/usr/bin/npm', 'run', 'start' ]
+2 info using npm@3.5.2
+3 info using node@v8.10.0
+4 verbose run-script [ 'prestart', 'start', 'poststart' ]
+5 info lifecycle blog@1.0.0~prestart: blog@1.0.0
+6 silly lifecycle blog@1.0.0~prestart: no script for prestart, continuing
+7 info lifecycle blog@1.0.0~start: blog@1.0.0
+8 verbose lifecycle blog@1.0.0~start: unsafe-perm in lifecycle true
+9 verbose lifecycle blog@1.0.0~start: PATH: /usr/share/npm/bin/node-gyp-bin:/home/bkn/WebstormProjects/blog/node_modules/.bin:/home/bkn/WebstormProjects/blog/node_modules/.bin:/home/bkn/.rvm/gems/ruby-2.6.3/bin:/home/bkn/.rvm/gems/ruby-2.6.3@global/bin:/home/bkn/.rvm/rubies/ruby-2.6.3/bin:/home/bkn/.local/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/usr/games:/usr/local/games:/snap/bin:/home/bkn/.rvm/bin:/home/bkn/.rvm/bin
+10 verbose lifecycle blog@1.0.0~start: CWD: /home/bkn/WebstormProjects/blog
+11 silly lifecycle blog@1.0.0~start: Args: [ '-c', 'cross-env PORT=3001 node index.js' ]
+12 silly lifecycle blog@1.0.0~start: Returned: code: 1  signal: null
+13 info lifecycle blog@1.0.0~start: Failed to exec start script
+14 verbose stack Error: blog@1.0.0 start: `cross-env PORT=3001 node index.js`
+14 verbose stack Exit status 1
+14 verbose stack     at EventEmitter.<anonymous> (/usr/share/npm/lib/utils/lifecycle.js:232:16)
+14 verbose stack     at emitTwo (events.js:126:13)
+14 verbose stack     at EventEmitter.emit (events.js:214:7)
+14 verbose stack     at ChildProcess.<anonymous> (/usr/share/npm/lib/utils/spawn.js:24:14)
+14 verbose stack     at emitTwo (events.js:126:13)
+14 verbose stack     at ChildProcess.emit (events.js:214:7)
+14 verbose stack     at maybeClose (internal/child_process.js:925:16)
+14 verbose stack     at Process.ChildProcess._handle.onexit (internal/child_process.js:209:5)
+15 verbose pkgid blog@1.0.0
+16 verbose cwd /home/bkn/WebstormProjects/blog
+17 error Linux 4.18.0-18-generic
+18 error argv "/usr/bin/node" "/usr/bin/npm" "run" "start"
+19 error node v8.10.0
+20 error npm  v3.5.2
+21 error code ELIFECYCLE
+22 error blog@1.0.0 start: `cross-env PORT=3001 node index.js`
+22 error Exit status 1
+23 error Failed at the blog@1.0.0 start script 'cross-env PORT=3001 node index.js'.
+23 error Make sure you have the latest version of node.js and npm installed.
+23 error If you do, this is most likely a problem with the blog package,
+23 error not with npm itself.
+23 error Tell the author that this fails on your system:
+23 error     cross-env PORT=3001 node index.js
+23 error You can get information on how to open an issue for this project with:
+23 error     npm bugs blog
+23 error Or if that isn't available, you can get their info via:
+23 error     npm owner ls blog
+23 error There is likely additional logging output above.
+24 verbose exit [ 1, true ]
