+++ package.json
@@ -4,13 +4,12 @@
   "description": "This is social blog.",
   "main": "index.js",
   "scripts": {
-    "start": "node index.js",
+    "start": "cross-env PORT=3001 node index.js",
     "dev": "nodemon index.js"
   },
   "author": "Vladislav Gorin (bkngg@yandex.ru)",
   "license": "ISC",
   "devDependencies": {
-    "body-parser": "^1.19.0",
     "eslint": "^5.16.0",
     "eslint-config-prettier": "^4.2.0",
     "eslint-plugin-node": "^9.0.0",
@@ -20,7 +19,11 @@
     "node": ">=9.0.0"
   },
   "dependencies": {
+    "body-parser": "^1.19.0",
+    "cross-env": "^5.2.0",
     "ejs": "^2.6.1",
-    "express": "^4.16.4"
+    "express": "^4.16.4",
+    "mongodb": "^3.2.3",
+    "mongoose": "^5.5.5"
   }
 }
